refactor!: merge load functions
