+++ src/expertsystem/io/_dict/__init__.py
@@ -1,11 +1,16 @@
-"""Serialization from and to a YAML recipe file."""
+"""Serialization from and to a `dict`."""
 
-import yaml
 
 from expertsystem.amplitude.model import AmplitudeModel
 from expertsystem.particle import Particle, ParticleCollection
 
-from . import _build, _dump
+from . import _build, _dump, _validate
+
+__all__ = [
+    "_build",
+    "_dump",
+    "_validate",
+]
 
 
 def asdict(instance: object) -> dict:
@@ -18,38 +23,3 @@ def asdict(instance: object) -> dict:
     raise NotImplementedError(
         f"No conversion for dict available for class {instance.__class__.__name__}"
     )
-
-
-class _IncreasedIndent(yaml.Dumper):
-    # pylint: disable=too-many-ancestors
-    def increase_indent(self, flow=False, indentless=False):  # type: ignore
-        return super().increase_indent(flow, False)
-
-    def write_line_break(self, data=None):  # type: ignore
-        """See https://stackoverflow.com/a/44284819."""
-        super().write_line_break(data)
-        if len(self.indents) == 1:
-            super().write_line_break()
-
-
-def load_amplitude_model(filename: str) -> AmplitudeModel:
-    with open(filename) as yaml_file:
-        definition = yaml.load(yaml_file, Loader=yaml.SafeLoader)
-    return _build.build_amplitude_model(definition)
-
-
-def load_particle_collection(filename: str) -> ParticleCollection:
-    with open(filename) as yaml_file:
-        definition = yaml.load(yaml_file, Loader=yaml.SafeLoader)
-    return _build.build_particle_collection(definition)
-
-
-def write(instance: object, filename: str) -> None:
-    with open(filename, "w") as yaml_file:
-        yaml.dump(
-            asdict(instance),
-            yaml_file,
-            sort_keys=False,
-            Dumper=_IncreasedIndent,
-            default_flow_style=False,
-        )
